requirements.txt
# WhyNotHireMe — Python dependencies
# Pin minor versions for reproducibility; patch versions left open.

# API framework
fastapi==0.111.*
uvicorn[standard]==0.29.*
python-multipart==0.0.*       # required for FastAPI file uploads

# PDF parsing
pdfplumber==0.11.*
PyMuPDF==1.24.*               # optional fallback; comment out if not needed

# NLP / embeddings
sentence-transformers==3.0.*
numpy>=1.26,<2.0

# HuggingFace (AI explanation layer)
transformers==4.41.*
torch>=2.2,<3.0               # CPU build is fine; use torch+cu121 for GPU

# Utilities
python-dotenv==1.0.*          # .env file support
pydantic==2.7.*               # data validation (FastAPI uses this internally)